NeighborlyAPI/requirements.txt
azure-functions==1.9.0
azure-cosmos==4.2.0
azure-cli-core==2.34.1
bson==0.5.10
pymongo==3.10.1
Flask==1.1.2
requests==2.23.0